+++ shpec/object_shpec.bash
@@ -1,0 +1,61 @@
+source import.bash
+
+shpec_helper_imports=(
+  initialize_shpec_helper
+  shpec_cwd
+  stop_on_error
+)
+eval "$(importa shpec-helper shpec_helper_imports)"
+initialize_shpec_helper
+stop_on_error=true
+stop_on_error
+
+source "$(shpec_cwd)"/../lib/rubsh.bash
+
+is_function () { declare -f "$1" >/dev/null 2>&1 ;}
+
+describe Object
+  it "is a function"
+    is_function Object
+    assert equal 0 $?
+  end
+
+  it "has no super"
+    [[ -z ${__superh[Object]-} ]]
+    assert equal 0 $?
+  end
+
+  describe class
+    it "Object is class Class"
+      Object .class
+      assert equal '"Class"' "$__"
+    end
+
+    it "Class is class Class"
+      Class .class
+      assert equal '"Class"' "$__"
+    end
+  end
+
+  describe methods
+    it "lists an array string of Object methods"
+      Object .methods
+      assert equal '([0]="ancestors" [1]="declare" [2]="instance_methods" [3]="new" [4]="superclass" [5]="class" [6]="methods")' "$__"
+    end
+
+    it "only lists Object methods defined on it when given false"
+      Object .methods false
+      assert equal '()' "$__"
+    end
+
+    it "lists an array string of Class methods"
+      Class .methods
+      assert equal '([0]="ancestors" [1]="declare" [2]="instance_methods" [3]="new" [4]="superclass" [5]="class" [6]="methods")' "$__"
+    end
+
+    it "only lists Class methods defined on it when given false"
+      Class .methods false
+      assert equal '()' "$__"
+    end
+  end
+end